requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
typer==0.9.0
pydantic==2.5.0
